Refactor LayoutHeader/Nav code * place nav popup handler code in Nav/Popup hook, useNavPopups
